Fix Marshalling for ruby 3
In more recent versions of ruby the behaviour of `YAML.load` has changed
to use safe loading by default. This change resotes the previous
behaviour by changing it to unsafe load. I considered using `safe_load`
with a list of permitted classes but I'm not sure what the exhaustive
list of possible classes would be.